+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.0.0 (not published)
 
 -   Migrate to Angular 10
+-   Enhance `<pxb-empty-state>` to allow ng-content as `title` or `description`.
 -   Add `isVisible` prop to the `<pxb-drawer-nav-item>`.
 
 ## v3.0.1